+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Typedoc for generating code documentation.
 #### Updated
 - Updated the Panel components to take in new props and using the Header from Reusable Components.
+- Updating the PlsIDForm to use an Id prop to render unique Ids on the page.
 
 ### v1.3.3
 #### Added