+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 #### Added
 - Typedoc for generating code documentation.
 #### Updated
-- Updated the Panel components to take in new props and using the Header from Reusable Components.
+- Updated the Panel components to take in new props and used the Header from Reusable Components.
+- Updated the PlsIDForm to use the `uuid` prop as an ID to its `Input` component for accessibility purposes.
 
 ### v1.3.3
 #### Added